requirements.txt
absl-py==0.12.0
appdirs==1.4.4
apt-xapian-index==0.49
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
blis==0.7.4
Brlapi==0.7.0
cachetools==4.2.2
catalogue==2.0.4
catfish==1.4.13
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
dataclasses==0.6
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
dill==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.6
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
fire==0.4.0
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gensim==4.0.1
gin-config==0.4.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.17.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
gpg===1.13.1-unknown
graphviz==0.16
grpcio==1.34.1
h5py==3.1.0
httplib2==0.19.1
idna==2.8
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.8.1
lightdm-gtk-greeter-settings==1.2.2
llvmlite==0.36.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
menulibre==2.2.1
mistune==0.8.4
monotonic==1.5
mugshot==0.4.2
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
neural-structured-learning==1.3.1
nltk==3.6.2
notebook==6.4.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandas-downcast==1.0.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pathy==0.5.2
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pooch==1.3.0
preshed==3.0.5
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.17.1
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.7.4
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-Levenshtein==0.12.2
python-slugify==5.0.2
pyttygif==0.2.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.95
sgt-launcher==0.2.5
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smart-open==3.0.0
SoundFile==0.10.3.post1
spacy==3.1.0
spacy-legacy==3.0.8
srsly==2.4.1
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.9.0
tensorflow-metadata==1.0.0
tensorflow-model-optimization==0.5.0
tensorflowjs==3.6.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
testresources==2.0.1
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.3.2
tflite-runtime==2.5.0
tflite-support==0.2.0
thinc==8.0.7
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xcffib==0.8.1
xkit==0.0.0
xx-ent-wiki-sm==3.1.0
zipp==3.4.1
